Add logging and task exit provision in workers, coordinator
Included logging functionality in worker and coordinator code to log to a file. Made changes to map, reduce, and the new "exit" tasks in the worker setup to return if they fail on submission. This helps with better error tracking and graceful task exit in the worker and coordinator functions in the MapReduce setup.